+++ backend/backend_main/cart/urls.py
@@ -1,0 +1,10 @@
+from django.urls import path, include
+# from rest_framework.routers import DefaultRouter
+
+from . import views
+
+# router.register(views.list_or_add, basename="cart")
+
+urlpatterns = [
+    path('', views.list_or_add, name='list-or-add' ),
+]
